remove legacy Task db type
