+++ packages/client/mutations/handlers/handleRemoveTasks.ts
@@ -2,8 +2,8 @@ import {RecordSourceSelectorProxy} from 'relay-runtime'
 import {handleRemoveReply} from '~/mutations/DeleteCommentMutation'
 import getDiscussionThreadConn from '~/mutations/connections/getDiscussionThreadConn'
 import {parseQueryParams} from '~/utils/useQueryParameterParser'
-import ITask from '../../../server/database/types/Task'
 import IUser from '../../../server/database/types/User'
+import {Task as ITask} from '../../../server/postgres/types/index.d'
 import safeRemoveNodeFromArray from '../../utils/relay/safeRemoveNodeFromArray'
 import safeRemoveNodeFromConn from '../../utils/relay/safeRemoveNodeFromConn'
 import getArchivedTasksConn from '../connections/getArchivedTasksConn'
