+++ packages/server/graphql/mutations/createTask.ts
@@ -9,11 +9,11 @@ import extractTextFromDraftString from '../../../client/utils/draftjs/extractTex
 import getTagsFromEntityMap from '../../../client/utils/draftjs/getTagsFromEntityMap'
 import getRethink from '../../database/rethinkDriver'
 import NotificationTaskInvolves from '../../database/types/NotificationTaskInvolves'
-import {TaskServiceEnum} from '../../database/types/Task'
 import generateUID from '../../generateUID'
 import updatePrevUsedRepoIntegrationsCache from '../../integrations/updatePrevUsedRepoIntegrationsCache'
 import getKysely from '../../postgres/getKysely'
 import {Task} from '../../postgres/types/index.d'
+import {TaskServiceEnum} from '../../postgres/types/TaskIntegration'
 import {analytics} from '../../utils/analytics/analytics'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import publish, {SubOptions} from '../../utils/publish'
