+++ packages/server/graphql/mutations/updatePokerScope.ts
@@ -3,9 +3,9 @@ import {Insertable} from 'kysely'
 import {SubscriptionChannel, Threshold} from 'parabol-client/types/constEnums'
 import {ESTIMATE_TASK_SORT_ORDER} from '../../../client/utils/constants'
 import EstimateStage from '../../database/types/EstimateStage'
-import {TaskServiceEnum} from '../../database/types/Task'
 import getKysely from '../../postgres/getKysely'
 import {Discussion} from '../../postgres/pg'
+import {TaskServiceEnum} from '../../postgres/types/TaskIntegration'
 import RedisLockQueue from '../../utils/RedisLockQueue'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import getPhase from '../../utils/getPhase'
