+++ packages/server/graphql/public/types/TaskEstimate.ts
@@ -1,0 +1,7 @@
+import {TaskEstimateResolvers} from '../resolverTypes'
+
+const TaskEstimate: TaskEstimateResolvers = {
+  label: ({label}) => label || ''
+}
+
+export default TaskEstimate
